requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-admin-thumbnails==0.2.5
django-ckeditor==6.0.0
django-currencies==0.10.1
django-js-asset==1.2.2
django-mptt==0.12.0
idna==2.10
jsonfield==2.1.1
Pillow==8.2.0
python-gettext==4.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
